# ARROW-12303: [JS] Use iterator instead of yield