--- a/js/src/util/bit.ts
+++ b/js/src/util/bit.ts
@@ -40,7 +40,7 @@ export function truncateBitmap(offset: number, length: number, bitmap: Uint8Arra
         // If the offset is a multiple of 8 bits, it's safe to slice the bitmap
         bytes.set(offset % 8 === 0 ? bitmap.subarray(offset >> 3) :
             // Otherwise iterate each bit from the offset and return a new one
-            packBools(iterateBits(bitmap, offset, length, null, getBool)).subarray(0, alignedSize));
+            packBools(new BitIterator(bitmap, offset, length, null, getBool)).subarray(0, alignedSize));
         return bytes;
     }
     return bitmap;
@@ -64,16 +64,52 @@ export function packBools(values: Iterable<any>) {
 }
 
 /** @ignore */
-export function* iterateBits<T>(bytes: Uint8Array, begin: number, length: number, context: any,
-                                get: (context: any, index: number, byte: number, bit: number) => T) {
-    let bit = begin % 8;
-    let byteIndex = begin >> 3;
-    let index = 0, remaining = length;
-    for (; remaining > 0; bit = 0) {
-        let byte = bytes[byteIndex++];
-        do {
-            yield get(context, index++, byte, bit);
-        } while (--remaining > 0 && ++bit < 8);
+// @ts-ignore
+export class BitIterator<T> implements IterableIterator<T> {
+    bit: number;
+    byte: number;
+    byteIndex: number;
+    index: number;
+
+    constructor(
+        private bytes: Uint8Array,
+        begin: number,
+        private length: number,
+        private context: any,
+        private get: (context: any, index: number, byte: number, bit: number) => T
+    ) {
+        this.bit = begin % 8;
+        this.byteIndex = begin >> 3;
+        this.byte = bytes[this.byteIndex];
+        this.index = 0;
+    }
+
+    next(): IteratorResult<T> {
+        if (this.index >= this.length) {
+            return {
+                done: true,
+                value: null
+            };
+        }
+
+        if (this.bit < 8) {
+            return {
+                value: this.get(this.context, this.index++, this.byte, this.bit++)
+            };
+        }
+
+        this.byteIndex++;
+        this.bit = 0;
+
+        if (this.byteIndex < this.bytes.length) {
+            this.byte = this.bytes[this.byteIndex];
+        }
+
+        return this.next();
+    }
+
+    [Symbol.iterator]() {
+        return this;
     }
 }
 
@@ -89,7 +125,7 @@ export function popcnt_bit_range(data: Uint8Array, lhs: number, rhs: number): nu
     // If the bit range is less than one byte, sum the 1 bits in the bit range
     if (rhs - lhs < 8) {
         let sum = 0;
-        for (const bit of iterateBits(data, lhs, rhs - lhs, data, getBit)) {
+        for (const bit of new BitIterator(data, lhs, rhs - lhs, data, getBit)) {
             sum += bit;
         }
         return sum;