@domoritz 2021-04-13
Relevant V8 issue: https://bugs.chromium.org/p/v8/issues/detail?id=7236